Fix aapt2_compile across cells
Summary:
When referencing resources in a different cell, aapt2 compile used to fail with "No such file or directory" because it would search within the `buck-out` folder of the current cell. This fixes that with the same methodology seen in https://github.com/facebook/buck/commit/e803c7f082

I'm not sure if this is the proper way to do this but my testing showed that it worked as expected with this change.

cc dreiss
Closes https://github.com/facebook/buck/pull/1383

Reviewed By: dreiss

Pulled By: dreiss

fbshipit-source-id: 313d93c