Fix project generator generating the whole action graph
Summary:
The previous code instantiated the action graph rooted at the binary node containing the linker-flags-with-macros field.
This new code instantiates only the action graph rooted at the macros' build target.

Test Plan:
ProjectGeneratorTests has a case that covers the behavior
I generated a project in the profiler and saw that it did not create the entire action graph

Reviewed By: dinhviethoa

fbshipit-source-id: 78ebb7f