CachingBuildEngine: buildLocally returns a Future
Summary:
I'm working towards allowing individual build rules to create their own `ListenableFuture`
when they need to be built locally. This commit is moving stuff around to lay the groundwork.

Due to the sensitive nature of this code, this change was made entirely with IntelliJ refactors
and two cut/pastes:
- Extract variable for `service.withDefaultAmounts(...)`
- Cut/paste that variable declaration outside the lambda
- Extract method for the body of the lambda, name the method `buildLocally`
- Cut/paste that method so it's above the other `buildLocally`
- Inline and delete the original `buildLocally`
- Inline and delete the variable extracted at the beginning

Test Plan: CI

fbshipit-source-id: d8f914c