CachingBuildEngine: Factor out BuildRule.buildLocally
Summary:
I'm working towards allowing individual build rules to create their own `ListenableFuture`
when they need to be built locally. This commit is moving stuff around to lay the groundwork.

Due to the sensitive nature of this code, this commit is just IntelliJ refactors and one manual
hoisting of variables out of a loop:
- Extract variables for the `BuildContext`, `StepRunner`, and `executionContext.withProcessExecutor`
- Cut/paste them all outside the for loop
- Extract method
- Move to BuildRule
- Rename a few things that were awkwardly auto-named
- Inline the variables that were extracted earlier

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 704b685