CachingBuildEngine: Inline some helpers into buildLocally
Summary:
I'm working towards allowing individual build rules to create their own `ListenableFuture`
when they need to be built locally. This temporarily uglifies the method in preparation for moving more
of it into `BuildRule`.

Due to the sensitive nature of this code, this change was made entirely with IntelliJ refactors
and one manual deletion:
- Inlined executeCommandsNowThatDepsAreBuilt into buildLocally
- Inlined wrapWithCancellationAndTracing into buildLocally
- Inlined the resulting function local
- Manually deleted the extra indirection left over from that

Test Plan: CI

fbshipit-source-id: 7015ae0