CachingBuildEngine: Create a new Scope object for buildLocally
Summary:
I'm working towards allowing individual build rules to create their own `ListenableFuture`
when they need to be built locally. In subsequent commits, we'll pass a `Supplier` of this scope in to
`BuildRule`, so that `BuildRule`s can create the scope when they
need it.

Unlike the previous commits in this stack, this one was entirely
done by hand.

Test Plan: CI

fbshipit-source-id: b14368a