CachingBuildEngine: buildLocally -> method object
Summary:
I'm working towards allowing individual build rules to create their own `ListenableFuture`
when they need to be built locally. There's enough state that needs keeping around that this gets easier
if it's an object.

Due to the sensitive nature of this code, this commit was created with
an IntelliJ refactor and some minor fixups:

- IntelliJ extract Method Object
- Cut/paste BuildLocallyScope to inner class of BuildLocally
- Manually delete ctor params of `BuildLocallyScope`

Test Plan: CI

fbshipit-source-id: 96ce818