CachingBuildEngine: Move timing to the scope and a success handler
Summary:
I'm working towards allowing individual build rules to create their own ListenableFuture
when they need to be built locally. This commit moves the timing logic for logging
out of the innermost future, so that it can remain inside `CachingBuildEngine`.

Due to the sensitive nature of this code, this commit was created largely with
IntelliJ refactors, but has a little more manual code that most of the commits
in this stack:

- Extract field for 'start'
- Extract method onStart
- Cut/paste onStart to the Scope constructor
- Extract method onSuccess
- Manually write the Futures.transform, put onSuccess and the return in it

Test Plan: CI

fbshipit-source-id: 27fdcf0