CachingBuildEngine: Allow BuildRule to provide its own ListenableFuture when built locally
Summary:
This concludes the refactor to allow individual build rules to create their own `ListenableFuture`
when they need to be built locally. This functionality will be used in subsequent commits to enable
a single build operation to produce the results of multiple rules.

Due to the sensitive nature of this code, I made heavy use of IntelliJ refactors, with a bit of
manual code:
- Extract method 'buildLocally'
- Extract parameter for each of the params to rule.buildLocally plus service
- Manually change the scope to come from Supplier<AutoCloseable>
- Inline rule.builtLocally
- Move builtLocally to BuildRule
- Remove unused parameter
- Added start if not started functionality

Test Plan: CI for now; as we use this code it'll get tested through the tests for its consumers

fbshipit-source-id: 3b026dc