Introduce BuildRuleResolver.computeIfAbsent and computeIfAbsentThrowing
Summary:
Many callsites use getRuleOptional + addToIndex in order to conditionally
create a rule if it does not exist. This diff adds computeIfAbsent and
computeIfAbsentThrowing which will fetch or call the supplier to create the
rule if the rule does not exist.

As a consequence, requireRule becomes a special case that computes the rule by
looking up the target node and running a transformer.

Subsequent diffs follow where I convert all uses of the old pattern to use
these methods.

This gets us closer to making action graph creation parallelizable.

Test Plan: CI

fbshipit-source-id: e7f3a37