Avoid recomputing transitive preprocessor deps for every source file
Summary:
A recent refactor to remove aggregated dependency rules for cxx compilation
caused preprocessor deps to be recomputed for every source file of a rule,
leading to massive build time increases. This code marks the function with
`Value.Lazy` so it is cached on first access.

Test Plan: Build a cxx rule with many dependencies and many files, see that action graph time is reduced.

Reviewed By: scottrice

fbshipit-source-id: 22c6535