Fix deadlock in FileSystemMap
Summary:
loader.load() can call back into other methods of the FileSystemMap.
Currently that can lead to a deadlock:
```
1. Entry.load() will lock entry without locking root, the load() call can call back into the map to
a method that requires acquiring the root lock
2. Several methods on FileSystemMap acquire the root lock followed by
locking an entry
```
Synchronizing all the Entry methods isn't really necessary. Accesses to subLevels already race. key is final. value can just be made
volatile.

This now synchronizes on the entry only around the loader.load() call
(so that we don't hash the same file multiple times).
This is safe as long as
```
1. no flow calls load() while holding the root lock
2. no flow calls load() followed by load() on a parent (i.e. loads need
  to propagate down).
```

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: dc444f5