CxxInferCaptureRulesAggregator has no dependencies
Summary:
This is a weird metadata-ish rule.

Currently it is always constructed with BuildRuleParams which always
throw on getBuildDeps. Accordingly it has never actually been depended
on in an ActionGraph.

Rather than propagating those always-erroring deps, don't even ask for
deps in the first place, and always act as if it has none.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: cc40f80