Avoid spurious exception when python & java race to delete tmp on ctrl-c
Summary:
We have at least one JVM shutdown hook that will try to delete
something in tmpdir (sqlite-jdbc tries to cleanup libsqlite.so), and that can
race with shutil.rmtree.  It's kind of clowny to throw on error to the user about
this, so catch ENOENT and just try again.

Test Plan: Ctrl-C some builds, observe the lack of errors

Reviewed By: dinhviethoa

fbshipit-source-id: 23358a4