new option to generate intellij project for targets under a folder
Summary:
With new option --intellij-project-root, buck project generates intellij project where buck targets under the specified folder are modules, and targets outside the folder are libraries.

IjProjectPaths is now an instance for IjProjectConfig.
All paths for IjModule and IjLibrary are converted to path relative to project root.

If --intellij-project-root is specified, the generated modules are not placed under a "module group".  This mimics behavior similar to imported Maven project.

Test Plan: unit test, and manual test

Reviewed By: styurin

fbshipit-source-id: 3e073c7