Jsr199Javac: Factor out CompilerBundle
Summary:
I'm working toward changing Java compilation so that we can
create the source ABI partway through compilation and allow
a module's dependents to start compiling while `javac` finishes
creating the full JAR for the module.

This commit introduces a `CompilerBundle` class to hold an instance
of the compiler and related objects and metadata. This will allow
multiple method calls on `Jsr199Javac` to all reference the same compiler.

Ideally someday we'd change the interface more drastically (to actually
proivde the `CompilerBundle` back to the caller rather than hiding it as
an implementation detail), but that's more involved than I want to bite
off just now.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 22e8e76