Different timeout for http read/write
Summary:
Within the investigation of cache artifacts that failed to be fetched on a bad network, we want to try setting different timeout for connection and receiving data on the network.
I'm introducing buck configuration to set them.

Test Plan:
- unit tests
- using network link conditioner on a Mac with delay at 300 ms for read and write,

With the configuration:
```
[cache]
    http_timeout_seconds=4
    http_read_timeout_seconds=4
    http_write_timeout_seconds=4
```

No error would show up

With the configuration:
```
[cache]
    http_timeout_seconds=4
    http_read_timeout_seconds=1
    http_write_timeout_seconds=1
```

Lots of fetch error would show.

fbshipit-source-id: 5ec6650