Make the at files smart
Summary:
This is a proposal that I made in form of code to make sure it works.

TLDR: This makes at files be optionally python scripts that could be in the buck root. The output is treated exactly as normal at files and the input are in the form of flavors that are passed to the script.

Test Plan: .

Reviewed By: yiding, dinhviethoa

fbshipit-source-id: c2c0e3c