Cache objects for out of process build
Summary:
Instead of creating and destroying these objects, we may cache them.
Also, tune the number of outer processes to be CORES/4: these processes are multithreaded themselves and according to my benchmarks,
the fastest mode for 8 core machine is when we have 2 outer processes for javac.
The results are:

   buck build buck -c java.location=IN_PROCESS
   takes 1m 5s

   buck build buck -c java.location=OUT_OF_PROCESS
   takes 1m 22s

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: 121312e