cxx: fix internal cache misses when calculating PCH hashes
Summary:
Since we use a `CxxToolFlags` object to index the PCH hash cache, the change to
using `SanitizedArg` in `CxxToolFlags`, which uses a new sanitizer function for
each arg, was causing the cache keys to never hit.  This diff makes sure we use
a consistent sanitizer function so that `SantizedArg`s return stable hash codes.

Test Plan: CI

Reviewed By: yiding

fbshipit-source-id: 1c951ac