More efficient hashCode for CxxSymlinkTreeHeaders
Summary:
This used to compute hash code for every field, which includes a name to path
map of all the headers. Since this object is derived from a HeaderSymlinkTree
BuildRule, we don't have to actually include the map as part of the
cache key.

This should have a small effect on action graph build time.

Test Plan:
CI

Ran `buck targets --show-output` with profiling to benchmark action graph creation.

In the test I ran, this change reduced time spent in
`CxxSymlinkTreeHeaders.hashCode` from 9 seconds to ~0.

Reviewed By: dinhviethoa

fbshipit-source-id: 342b641