Stop reading all source files in memory for DistBuildState.
Summary:
We used to read file contents as we record the files to be materialized,
regardless of whether or not it's already present in the CAS. This diff changes
this behaviour to only read the missing files in memory. Note that this only
reduces the memory footprint of the build, not the time, because we anyways need
to read the file to compute the SHA1.

Test Plan: CI

Reviewed By: alisdair04

fbshipit-source-id: eee46b3