Add option to override server-side config with a custom buckconfig file.
Summary: This diff adds a config option, `stampede.server_buckconfig_override` that allows users to specify the path to a .buckconfig file which would override the build-slave's buckconfig. Note that this buckconfig is as limited in its power as the local buckconfig that is sent. Because both of these configs are only applied during runtime, after we fetch the distributed build state from the server, these cannot affect any config settings which are taken care of in Main.java.

Test Plan: CI

Reviewed By: alisdair04, ruibm

fbshipit-source-id: c145fe1