Fix a bug in 'roots' selection
Summary:
We build a set of symlinks to source folders, or 'roots'. Any folder with a `BUCK`
file that contains source is a root candidate: Currently, we choose between a given
root candidate and any of its ancestors by choosing the one that results in the
fewest `<excludedFolder/>` tags. But! `a/b/c/d` may 'want' `a/b/c` while `a/b/e` may 'want'
`a/b`, which would be the directory containing `a/b/c`. The solution here is to add a
second pass, where we first generate a set of candidate roots, then replace each
candidate with its highest ancestor from the candidate set, if any. This IS a
second geometric pass, but doesn't seem so bad, in practice, because we limit the
size of one of the terms by skipping ahead whenever multiple source folders use the
same root.

Test Plan: `buck test //test/com/facebook/buck/ide/intellij/projectview && buck test //test/com/facebook/buck/ide/intellij`

Reviewed By: asp2insp

fbshipit-source-id: 0754751