Jsr199Javac: Push JarBuilder into CompilerBundle
Summary:
This is part of a series of commits intended to rearrange `Jsr199Javac` to
support `buildSourceAbiJar` and `buildClasses` being called on the same instance.

This commit moves the `JarBuilder` creation logic from `buildWithClasspath` into
`CompilerBundle`, so that it can be available to both `buildSourceAbiJar` and `buildClasses`.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 2b4749c