Jsr199Javac: Move closeables to a single array in CompilerBundle
Summary:
This is part of a series of commits intended to rearrange `Jsr199Javac` to
support `buildSourceAbiJar` and `buildClasses` being called on the same instance.

This commit replaces various fields that only existed so `close()` could be called
on them with a single array of `AutoCloseable`. Just simplifies things a bit.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: c01f01f