Jsr199Javac: Transform CompilerBundle into the Invocation
Summary:
This is part of a series of commits intended to rearrange `Jsr199Javac` to
support `buildSourceAbiJar` and `buildClasses` being called on the same instance.

The previous commits in this series have enabled us to make `CompilerBundle` be the
`Javac.Invocation` implementation for `Jsr199Javac`, which in turn will make it easier
to share a `javac` instance between `buildSourceAbiJar` and `buildClasses`. This commit does
so, renaming the class appropriately.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 00967ef