Jsr199JavacInvocation: Inline buildWithClasspath into buildClasses
Summary:
This is part of a series of commits intended to rearrange `Jsr199Javac` to
support `buildSourceAbiJar` and `buildClasses` being called on the same instance.

This commit inlines the build logic from `buildWithClasspath` into `buildClasses`.
This was mostly done with IntelliJ refactors with a couple manual tweaks before and
after.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: b632dab