Allow source ABI generation for all targets
Summary:
Previously it was assumed that source ABI would eventually always
be used without deps, which therefore required annotation processors
to be handled specially. Since I'm working toward changing Java
compilation so that we can create the source ABI partway through
compilation, we need to allow source ABIs to be generated for any target
and with all APs running un-sandboxed. The sandboxing code will come back
when we return to the source-abi-without-deps work.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 2369f80