Remove support for output-as-you-go shell steps
Summary:
With --verbose 3, ndk_library rules set this option in order to show the
commands make was running (because v=3 is supposed to show commands).
However, that dumps junk into stdout in a nondeterministic way (because
ndk_library rules are often fetched from cache).  Because ndk_library
is deprecated, just drop this feature.  We'll still dump the full output
to the log or console.

Test Plan: CI

fbshipit-source-id: c254a63