Fix buck-out/non-buck-out file hash cache handling
Summary:
Currently, the non-buck-out file hash cache is handling buck-out file
hashes. This is probably because ProjectFilesystem's handling of whether
or not it ignores buck-out is rather unintuitive.

This changes the isIgnored handling of the hash caches to be more
explicit and depend less on the ProjectFilesystem implementation.

Test Plan: CI, added tests for default and watched file hash caches

Reviewed By: andrewjcg

fbshipit-source-id: 49f953e