Jsr199JavacInvocation: Implement and use buildSourceAbiJar
Summary:
I'm working toward changing Java compilation so that we can
create the source ABI partway through compilation and allow
a module's dependents to start compiling while `javac` finishes
creating the full JAR for the module.

This commit implements the separate method on `Javac.Invocation`
for creating just the source ABI. In future this will allow the `JavacStep`
to create the ABI, then set a `Future`, then resume compilation. For now,
however, we just use it for the existing source ABI creation rules.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: ff5e372