DefaultJavaLibrary: Pass in ZipArchiveDependencySupplier rather than creating it
Summary:
I'm working toward changing Java compilation so that we can
create the source ABI partway through compilation and allow
a module's dependents to start compiling while `javac` finishes
creating the full JAR for the module.

With that available, library compilation might be started by either the
`CalculateAbiFromSource` rule or the `DefaultJavaLibrary` one, depending
on whether source ABIs are enabled. If we left it up to the rule to create
the supplier (and, say, forgot to do so), dep file keys would cease to function.
Factoring it out prevents us from getting that wrong. (Or, more accurately,
is the result of having gotten it wrong and debugged it for longer than I should have.)

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 7953a3d