Stop catching IOException and InterruptedException in Step#execute.
Summary:
A lot of steps catch IOException or InterruptedException to log errors.

This should be unnecessary as there is a lot of error handling code in `StepFailedException` for these cases.

If additional context is needed in the error message the step's description can be made more exhaustive.

Test Plan: CI

Reviewed By: aiked

fbshipit-source-id: dabb718