Make FilteredResourcesProvider.getResDirectories return SourcePaths
Summary:
This actually makes a lot of things simpler.  getResDirectories can be
called at any time now.  In getBuildSteps where we used to compute the
list of output dirs, I added some checks to make sure we have the exact
right number.

I had to remove two tests that were depending on the ability to lie
about what resource paths a given FilteredResourcesProvider had.  One is
not of high value.  The other can be replaced with an integration test
if necessary.

Test Plan: CI

fbshipit-source-id: 330245a