Sort reference maps based on key ids
Summary:
The complex references (maps) in arsc should be sorted by the ids of the
keys. If we transform the keys, we need to update the order of the map.
This is similar to how they are sorted in ResourcesXml (though there the
id needs to be indirectly looked up instead of available directly in the
data).

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 2feb935