Factor out static helper for BuildRule.buildLocally
Summary:
I'm working toward changing Java compilation so that we can
create the source ABI partway through compilation and allow
a module's dependents to start compiling while javac finishes
creating the full JAR for the module.

In so doing, I need to run steps that belong to more than one rule,
and thus having the implementation in the body of `BuildRule.buildLocally`
itself is inconvenient.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 9af1ae0