Factor out common portions of CalculateAbiFromSource impl
Summary:
I'm working toward changing Java compilation so that we can
create the source ABI partway through compilation and allow
a module's dependents to start compiling while `javac` finishes
creating the full JAR for the module.

In this world, both `CalculateAbiFromSource` and `DefaultJavaLibrary`
need to know how to build the full library, and `DefaultJavaLibrary`
needs to be able to get the one built by the other rule. We'll accomplish
this by sharing a common worker object between the two rules. This commit
introduces that worker class, and moves the implementation of
`CalculateAbiFromSource` into it.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 13865cd