Factor out common portions of DefaultJavaLibrary impl
Summary:
I'm working toward changing Java compilation so that we can
create the source ABI partway through compilation and allow
a module's dependents to start compiling while `javac` finishes
creating the full JAR for the module.

Continuing from the previous commit in this series, this commit
moves the common elements of `DefaultJavaLibrary` into the worker
class.

The majority of the files changed in this commit were to remove
constructor parameters and fields that become unused as a result
of the change. The meat of the change is in `DefaultJavaLibrary`
and `JavaAbiAndLibraryWorker`.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 2aa6fa6