JavaAbiAndLibraryWorker: Take responsibility for recording artifacts
Summary:
I'm working toward changing Java compilation so that we can
create the source ABI partway through compilation and allow
a module's dependents to start compiling while javac finishes
creating the full JAR for the module.

In this world, the output of one rule might be built as part
of building another. This commit makes sure the artifacts are
all reported as part of the correct rules.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: 8eb66cb