Generate source ABIs on the way to outputting the full jar
Summary:
This commit enables source ABIs to be generated as a side effect of
generating a full jar. Further, it starts building dependent rules
as soon as the source ABI is available.

The way this works is that the library and abi rules share a worker
object, and get their `buildLocally` futures from there.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: f6212e6