Refer to targets in SQLite by their cell-relative name
Summary:
Using a fully-qualified target name as a DB key actually causes
ambiguity when targets are referred to by both a root-cell build and a
cross-cell build.  This can lead to the metadata db getting out of sync with the
filesystem, most notably when RECORDED_PATHS is deleted (see the test case).

Fixes #1363

Test Plan: new integration test

Reviewed By: asp2insp

fbshipit-source-id: 549c9c9