Introduce a way for apple_resources to reexport transitive resources of another rule
Summary:
Currently, when collecting dependencies to include in a bundle, a traversal is
done which proceeds through library dependencies to find apple_resource rules
that will be included in the bundle. This traversal stops upon hitting any rule
which is not within the small set of standard rules.

This diff introduces a new field on `apple_resource`, `resources_from_deps`,
which has the effect of re-exporting all the transitive resource dependencies
of a different target.  This allows for resource rules to directly depend on
other resources, or on libraries or bundles that has resource dependencies.

This is useful to continue traversal for resources where otherwise the default
traversal mechanism stops, such if a library is transformed via a genrule.

Alternative mechanisms considered includes:

1. Adding a field to genrule itself to allow genrules to selective stop the
   resource traversal. I don't like this method since it doesn't really scale
   for other resource collection types.

2. Reworking how resource collection is done by apple_bundle to allow for a
   more flexible specification. This method is nice to have in the long term,
   but its interface will require more care to develop than I have time for
   right now.

Test Plan: Added an integration test to verify such a dependency is included.

Reviewed By: bertmaher

fbshipit-source-id: 6f10815