BuildRuleParams no longer has a ProjectFilesystem
Summary:
Thread one through in parallel where one is needed.

There are a lot of places where semi-invalid BuildRuleParams are passed
around, or where a lot more information is passed around than is
actually needed. This is a step in reducing the amount that we do that.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 7b48372