testVersionUniverseChangeCausesMiss actually tests with VersionUniverses
Summary:
Before, it only tested that new TargetNodes were constructed, and not
that they were actually different.

I have an upcoming change which makes TargetNodes use semantic equality,
rather than identity equality, which breaks this test. Flesh out the
test so that the VersionUniverse actually makes a difference in the
constructed VersionedTargetGraph, so that they're not equal to each
other.

Test Plan: CI

Reviewed By: Coneko

fbshipit-source-id: 956aad9