Use the remote_sandbox root cell for everything.
Summary:
We were using the remote_sandbox cell at most of the places, but the
normal slave build cell for the Build object, target resolution at a few places,
and AndroidDirResolver. There does not seem to be a good reason for doing this,
and was probably erroneous. This diff gets rid of the confusion.

Test Plan: CI

Reviewed By: alisdair04

fbshipit-source-id: 723663f