Add HasInstallHelpers interface
Summary:
This interface allows a buildrule to return a list of targets that
should be built when installing that rule.

For example, if the rule for `//:foo` returns `//:foo#install-helper1,
//:bar`, then `buck install //:foo` will build `//:foo#install-helper1`
and `//:bar`. It is the rule's responsibility to ensure that these
targets are in the action graph (e.g. by always putting them there even
in non-install contexts).

This approach ensures that the target and action graph are the same
between build and install (though ios bundles already violate this with
a unique adhoc mechanism).

Test Plan: CI

Reviewed By: bertmaher

fbshipit-source-id: 7b9c234