PrefixMapDebugPathSanitizer: don't pad with `/`s
Summary:
To support post-mortem debug path sanitization, the `MungingDebugPathSanitizer`
needs to make sure the embedded compilation dir gets padded to a fixed size with
trailing `/`s.  However, the `PrefixMapDebugPathSanitizer` relies on a different
strategy and doesn't require this eye-sore.

This diff refactors the sanitizers to move munging-specific stuff, including
compilation dir padding, into the `MungingDebugPathSanitizer` class so that
the `PrefixMapDebugPathSanitizer` doesn't use it.

Test Plan: CI

fbshipit-source-id: 0b9cb7e