Don't discard IOExceptions
Summary:
This code is using a strange pattern where it discards the IOException
and only propagates the cause. This was (I expect) unintentional.

These were the only places in the codebase (that I could find with some
simple grepping) that were doing this (there are a bunch of cases
unwrapping getCause() from ExecutionException and similar that are meant
for that, but none doing it with IOException or other such types).

Test Plan: CI

Reviewed By: yiding

fbshipit-source-id: dadcbbc