Added exopackage installer flag to just kill single process by name
Summary: When we edit code that lives in a background process, we don't need to do a full coldstart to pick up the changes, we just need to kill the background process. This change allows buck to kill a specific processes by name following an exo-install rather than killing the entire app.

Test Plan: run `buck install -p <package> <app>` and see the background process die

Reviewed By: dreiss

fbshipit-source-id: eeddb7b