Don't fail builds on use of deprecated things
Summary:
We have fatal linters available in PMD if we really want to avoid using
a type (e.g. `PrintWriter`). Deprecation gives a nice nudge to people
in the IDE when they try to use a class or method that we're trying to
get rid of. Having to go through and mark all the usages with `SuppressWarnings`
when you want to deprecate something discourages it.

Test Plan: CI

Reviewed By: illicitonion

fbshipit-source-id: 5e1906d