AddsToRuleKey 3/3: Enable any class to use @AddToRuleKey
Summary:
This commit implements the actual change of enabling `AddToRuleKey` to be used from
any class that implements `AddsToRuleKey`. This is now the preferred method of adding
things to rule keys; accordingly `RuleKeyAppendable` is now marked deprecated.

I deliberately did not have `BuildRule` implement `AddsToRuleKey`, as `BuildRule`s are
treated slightly differently in some places, and it would be easy to accidentally get
that wrong in future.

Test Plan: Updated unit test

Reviewed By: dreiss

fbshipit-source-id: 17e4b81