Move all single IDevice things to RealAndroidDevice
Summary:
Moves a bunch of functions from AdbHelper to RealAndroidDevice. Each of
these functions took a single IDevice as a parameter and then did some
operation on it. Instead, they are now all methods on RealAndroidDevice
that operate on the underlying IDevice.

These functions were mostly just used from each other and stuff that was
already in RealAndroidDevice, and just a few were called from AdbHelper.

Moves the tests of these functions in RealAndroidDeviceTest.

Test Plan: CI

Reviewed By: bertmaher

fbshipit-source-id: dd9316f