Add InstallTrigger
Summary:
This adds a SourcePath that will have different contents on every
`buck install` invocation (and isn't available in non-install commands). This
can be used to construct a BuildRule that will always be run. This will allow
the work of installing to be done concurrently with all the other build work
(and allows such work to leverage all the same techniques for logging/avoiding
 work/tracing/etc that buildrules have).

Test Plan: CI, added tests

Reviewed By: yiding

fbshipit-source-id: 473ac05