Memoize AdbHelper.getDevices()
Summary:
This just memoizes the devices list. If this list is changing in between
calls (like someone adding/removing a device) our current behavior is
going to be messed up. Now all commands on the same AdbHelper will
operate over the same list of devices in the same order (and should fail
if a device is removed and ignore devices that are added).

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: c10dc0c