Evaluate target literals at parsetime
Summary:
We want to immediately evaluate targets if the type of the TargetEvaluator is not LAZY. In order to store the evaluated QueryTargets, I create a TargetSetExpression, which replaces both TargetLiteral and SetExpression when the evaluator's type is IMMEDIATE.

With this change, it becomes apparent that QueryExpression.collectTargetPatterns has two distinct uses: for preloading those patterns before evaluating and for extracting build targets. The first use is unaffected, but we must now include pre-evaluated targets in the second. I replaced it with a new QueryExpression.Visitor that includes both evaluated TargetLiterals and pre-evaluated TargetSetExpressions.

Test Plan: Added tests of new visitor, modified mocking in existing tests to track calls to getTargetEvaluator()

Reviewed By: asp2insp

fbshipit-source-id: a1cc17e