Create QueryEvaluator for controlling evaluation of subqueries
Summary:
In order to cache subexpressions in queries as we evaluate them, we need some way to take control of evaluation back from the QueryExpression. By threading a QueryEvaluator through evaluation, we can do this at any node of the expression.

For now, use a NoopQueryEvaluator that immediately passes back to the expression for evaluation.

Test Plan: Functionality unchanged, ran `buck build` successfully

Reviewed By: illicitonion, asp2insp

fbshipit-source-id: 540685d