Gracefully fail when undefined symbols are present
Summary:
If the source ABI is compiled before the full jar, and undefined symbols
are present, Buck would crash because such symbols are represented by
`ErrorType`s and there were a couple places where we threw because support
for those wasn't implemented. This commit adds reasonable handling for those cases,
keeping in mind that the existence of an `ErrorType` during ABI generation
means that the compiler has already issued errors and the build will
be ending in failure.

Test Plan: New integration test

fbshipit-source-id: ea33e47