(IjProject) Add an ability to generate AndroidManifest.xml
Summary:
Support for modules with Android resources that has non-default package names is broken right now. If the package name doesn't match the package name in manifest located in `android_res/AndroidManifest.xml` and the target doesn't point to a custom manifest, the resources from such modules are not recognized.

The solution is to generate empty manifests with package names that match package names specified in targets.

Test Plan: Unit tests and local testing

Reviewed By: asp2insp

fbshipit-source-id: 68d9c35