DefaultJavaLibraryBuilder: Bring sanity to classesToRemoveFromJar
Summary:
`classesToRemoveFromJar` existed both in `JavacOptions` and as a separate field in `DefaultJavaLibrary`.
Only the one in `JavacOptions` was doing anything. It doesn't look like anything was broken by this (anything
that set the field did so by copying the value from `JavacOptions`), but it certainly was confusing.

This commit makes the field the definitive source.

Test Plan: Existing `DefaultJavaLibraryIntegrationTest` and `CalculateAbiFromSourceIntegrationTest`.

Reviewed By: dreiss

fbshipit-source-id: 6ad49d6