Add AndroidBinaryInstallIntegrationTest
Summary:
This adds an integration test for `buck install //:some_android_binary`.

Sets up a simple project that can be configured by editing a json file
and running a python script. This allows (very limited, but easy) modification of:
1. multiple java file data
2. multiple cxx file data
3. assets, resources, java_resources data
4. main apk java data
5. main apk resources data
6. manifest data
7. java dependencies
8. cxx dependencies
9. resource dependencies
10. exopackage on/off
11. package name

Then add a whole bunch of tests (almost exactly the same tests as in
ExopackageInstallerIntegrationTest) that installs and reinstalls
after changes properly update the state on the device. By making the
AdbHelper in the ExecutionContext return a controlled list of
AndroidDevices, this test can have `buck install` operate on
TestAndroidDevice (that writes apks/files to a directory on the host
instead of a real device).

This doesn't yet verify that the files on the "device" actually contain
the correct data, just that the correct number apks/files are pushed.

Test Plan: CI

Reviewed By: dreiss

fbshipit-source-id: d95ff1a