Expose hash cache engine's getIfPresent()
Summary:
A bunch of test's are using willGet() as an indication of whether or not
the cache contains an entry for a path. This is odd and depends on the
fact that willGet() call ProjectFilesystem.exists() on the path.

Some other tests were digging into the engine and calling its
getIfPresent(). This is also odd as that's just some internal detail of
the cache.

Instead, just expose the underlying engine's getIfPresent() and use that
in tests.

Test Plan: CI

Reviewed By: bertmaher

fbshipit-source-id: f2d972f