Make CachingBuildEngine track async callbacks directly
Summary:
The async callbacks queue is used for tracking work that should be finished,
but doesn't need to be done immediately. This is mostly just uploading
to cache and computing some extra information for logging.

Previously, this was done by creating a queue in build() and adding all
discovered tasks to that queue and then blocking the returned future on
those results. That doesn't really work as expected when there are
multiple concurrent build() calls.

This changes CachingBuildEngine to instead maintain the queue as a
member and then ensure all the tasks are finished in .close().

Test Plan: CI

Reviewed By: yiding

fbshipit-source-id: 82f33ac