Only one thread at a time should be loading the manifest
Summary:
I forgot to put this in place for the trie; a wip branch has it but not main.
The treemaps in the trie don't like it when multiple threads are modifying
them, nor do we need to pay the cost of parsing 1e6+ entries twice or more.

Test Plan: CI

Reviewed By: marcinkosiba

fbshipit-source-id: 05e534a